[225] Update python version defaulting
Updating python version overrides for an existing python project. If the
pythonVersion config is set then it will still be the desired version
but if it is not then we will delegate the python version to the
existing projects version. This can be updated by the new config
defaultPythonStrategy.
Created example project for defaultPythonStrategy

Signed-off-by: woods-cpointe <cwoods@cpointe-inc.com>