[#423] pull repository handling into class
Pulls logic for constructing the package index and publishing URLs for a
given python repository into a dedicated class.  Also moves some shared
parameters for controlling repos into the parent Mojo.  There are some
places where the new class could have been used, but to limit the scope
of this change, that will be left for future work.